winit: Rename xdg_color_scheme module to xdg_desktop_settings
The module handles color scheme, accent color, and will soon handle
cursor blink settings. The new name better reflects its scope.

Also rename the cfg alias from use_winit_theme to xdg_desktop_settings
(with inverted polarity) for consistency with the module name.
