+++ internal/backends/winit/build.rs
@@ -11,7 +11,7 @@ fn main() {
        enable_femtovg_renderer: { any(feature = "renderer-femtovg", feature = "renderer-femtovg-wgpu") },
        enable_accesskit: { all(feature = "accessibility", not(target_arch = "wasm32")) },
        supports_opengl: { all(any(feature = "renderer-skia-opengl", feature = "renderer-femtovg"), not(ios_and_friends)) },
-       use_winit_theme: { any(target_family = "windows", target_vendor = "apple", target_arch = "wasm32", target_os = "android") },
+       xdg_desktop_settings: { not(any(target_family = "windows", target_vendor = "apple", target_arch = "wasm32", target_os = "android")) },
        muda: { all(feature = "muda", any(target_os = "windows", target_os = "macos")) },
     }
     println!("cargo:rustc-check-cfg=cfg(slint_nightly_test)");
