+++ internal/backends/winit/lib.rs
@@ -95,8 +95,8 @@ mod renderer {
 mod accesskit;
 #[cfg(muda)]
 mod muda;
-#[cfg(not(use_winit_theme))]
-mod xdg_color_scheme;
+#[cfg(xdg_desktop_settings)]
+mod xdg_desktop_settings;
 
 #[cfg(target_arch = "wasm32")]
 pub(crate) mod wasm_input_helper;
